Split finality tests
