+++ packages/spec-test-runner/test/spec/finality/finality_1_fast.test.ts
@@ -1,0 +1,5 @@
+import {runFinalityTests1} from "./finality_fast";
+
+describe("finality fast 1", () => {
+  runFinalityTests1();
+});
