+++ packages/spec-test-runner/test/spec/finality/finality_2_fast.test.ts
@@ -1,0 +1,5 @@
+import {runFinalityTests2} from "./finality_fast";
+
+describe("finality fast 2", () => {
+  runFinalityTests2();
+});
